AdminController is now transformed into 4 Controllers and then deleted.
