+++ HoneyZoneMvc/Areas/Admin/Controllers/HomeController.cs
@@ -2,8 +2,7 @@
 
 namespace HoneyZoneMvc.Areas.Admin.Controllers
 {
-    [Area("Admin")]
-    public class HomeController : Controller
+    public class HomeController : BaseAdminController
     {
         public IActionResult Index()
         {
