+++ HoneyZoneMvc/Areas/Admin/Views/Shared/_OrdersTablePartial.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<OrderInfoViewModel>
+﻿@model IEnumerable<OrderAdminViewModel>
 @{
 }
 <table class="table mb-10 mt-3 table-striped table-light table-bordered border-warning table-hover table-sm datatable" id="orders">
